refactor(secure_rng): split 1137-line monolith into cohesive modules
Replace src/secure_rng/secure_rng.c with single-responsibility units
(per soul.md §10):

  secure_internal.h  - private shared helpers + internal constants/macros
  secure_init.c     - locks, entropy/health helpers, config, init family, reset, reseed
  secure_generate.c - bytes/uint64/uint32/double/range32/range64 generation
  secure_stats.c    - stats/state/print, error callback, mode control, version
  secure_selftest.c - built-in self-test

Verified: libs build clean (-Wall -Wextra, zero warnings); secure_rng_test
passes 23/23 (identical output). Pure refactor, no behavior change.

The Makefile builds src/secure_rng/*.c via wildcard, so new modules link
with no Makefile edit. Shared helpers (lock_*/collect_tested_entropy/
reseed_needed/run_bell_certification) are declared in secure_internal.h and
defined non-static in secure_init.c so generation/stats modules can call them.
